Add people module handler
