+++ modules/book_handle.rb
